Endpoint moved to personenkontext (just like BE) and added new spec
