+++ src/stores/PersonStore.spec.ts
@@ -183,61 +183,4 @@ describe('PersonStore', () => {
       expect(personStore.loading).toBe(false);
     });
   });
-
-  describe('createPerson', () => {
-    it('should create a Person', async () => {
-      const mockPerson: PersonendatensatzResponse = {
-        person: {
-          id: '9876',
-          name: {
-            familienname: 'Cena',
-            vorname: 'Randy',
-          },
-          referrer: 'rcena',
-        },
-      } as PersonendatensatzResponse;
-
-      const mockResponse: PersonendatensatzResponse = mockPerson;
-
-      mockadapter.onPost('/api/dbiam/personen').replyOnce(201, mockResponse);
-      const createPersonPromise: Promise<PersonendatensatzResponse> = personStore.createPersonWithKontext({
-        familienname: 'Cena',
-        vorname: 'Randy',
-        organisationId: '1234',
-        rolleId: '5678',
-      });
-      expect(personStore.loading).toBe(true);
-      const createdPerson: PersonendatensatzResponse = await createPersonPromise;
-      expect(createdPerson).toEqual(mockPerson);
-      expect(personStore.loading).toBe(false);
-    });
-
-    it('should handle string error', async () => {
-      mockadapter.onPost('/api/dbiam/personen').replyOnce(500, 'some error');
-      const createPersonPromise: Promise<PersonendatensatzResponse> = personStore.createPersonWithKontext({
-        familienname: 'Copeland',
-        vorname: 'Christian',
-        organisationId: '',
-        rolleId: '5678',
-      });
-      expect(personStore.loading).toBe(true);
-      await rejects(createPersonPromise);
-      expect(personStore.errorCode).toEqual('UNSPECIFIED_ERROR');
-      expect(personStore.loading).toBe(false);
-    });
-
-    it('should handle error code', async () => {
-      mockadapter.onPost('/api/dbiam/personen').replyOnce(500, { code: 'some mock server error' });
-      const createPersonPromise: Promise<PersonendatensatzResponse> = personStore.createPersonWithKontext({
-        familienname: 'Copeland',
-        vorname: 'Christian',
-        organisationId: '1234',
-        rolleId: '',
-      });
-      expect(personStore.loading).toBe(true);
-      await rejects(createPersonPromise);
-      expect(personStore.errorCode).toEqual('some mock server error');
-      expect(personStore.loading).toBe(false);
-    });
-  });
 });
